Revert "refactor(hooks): move hooks to root directory (#132)"
This reverts commit 69a6960f775e279983b0de16894cfa52d1b5d8e5.
