+++ frontend/src/app/(main)/chat/MainLayout.tsx
@@ -1,5 +1,6 @@
 'use client';
 import React, { useEffect, useState } from 'react';
+import { useAuthContext } from '@/app/providers/AuthProvider';
 import { useRouter } from 'next/navigation';
 import { ResizablePanel, ResizablePanelGroup } from '@/components/ui/resizable';
 import { SidebarProvider } from '@/components/ui/sidebar';
@@ -8,9 +9,8 @@ import { ChatSideBar } from '@/components/sidebar';
 import ProjectModal from '@/components/project-modal';
 import { useQuery } from '@apollo/client';
 import { ProjectProvider } from '@/components/code-engine/project-context';
-import { useChatList } from '@/hooks/useChatList';
+import { useChatList } from '@/app/hooks/useChatList';
 import { GET_USER_PROJECTS } from '@/graphql/request';
-import { useAuthContext } from '@/providers/AuthProvider';
 
 export default function MainLayout({
   children,
