+++ frontend/src/app/hooks/useSpeechRecognition.ts
