+++ frontend/src/app/providers/DynamicApolloProvider.tsx
