+++ frontend/src/components/AuthChoiceModal.tsx
