+++ frontend/src/components/fileSidebar.tsx
